Add Primitive data type

+++ Assignment 2/src/model/Person.java
@@ -20,7 +20,7 @@ public class Person {
                 public String street;
                 public String city;
                 public long phone;
-                public int unit; 
+                public short unit; 
                 public int zip;
                 public String state;
 
@@ -45,11 +45,11 @@ public class Person {
                     return phone;
                 }
 
-                public int getUnit() {
+                public short getUnit() {
                     return unit;
                 }
 
-                public void setUnit(int unit) {
+                public void setUnit(short unit) {
                     this.unit = unit;
                 }
 
@@ -170,7 +170,7 @@ public class Person {
         return homeAddress;
     }
 
-    public void setHomeAddress(String street, String state, int zip, String city, int unit, long phone) {
+    public void setHomeAddress(String street, String state, int zip, String city, short unit, long phone) {
         this.homeAddress.setStreet(street);
         this.homeAddress.setCity(city);
         this.homeAddress.setState(state);
@@ -183,7 +183,7 @@ public class Person {
         return workAddress;
     }
 
-    public void setWorkAddress(String street, String state, int zip, String city, int unit, long phone) {
+    public void setWorkAddress(String street, String state, int zip, String city, short unit, long phone) {
         this.workAddress.setStreet(street);
         this.workAddress.setCity(city);
         this.workAddress.setState(state);
